+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+- client: add `GeyserGrpcBuilder` ([#307](https://github.com/rpcpool/yellowstone-grpc/pull/307))
+
 ## 2024-03-20
 
 - yellowstone-grpc-client-1.14.0+solana.1.18.7